+++ src/frame/types.rs
@@ -5,11 +5,11 @@ pub(crate) type Bit = bool;
 
 pub(crate) type Word = u16;
 
-pub type Quantity = u16;
+pub type Quantity = u32;
 
 pub(crate) const REQUEST_BYTE_LAST_LEN: usize = 10;
 
-pub(crate) const LIMIT: u16 = 900;
+pub(crate) const LIMIT: u32 = 900;
 
 // 定义一个全局常量来表示超时时间
 pub(crate) const TIMEOUT_DURATION: Duration = Duration::from_millis(100);
